Extract duplicate code in ChunkPlatform between 2.13 and 3
