+++ core/shared/src/main/scala-3/fs2/ChunkPlatform.scala
@@ -27,25 +27,7 @@ import scala.collection.immutable.ArraySeq
 import scala.collection.immutable
 import scala.reflect.ClassTag
 
-private[fs2] trait ChunkPlatform[+O] { self: Chunk[O] =>
-
-  def toArraySeq[O2 >: O: ClassTag]: ArraySeq[O2] = {
-    val array: Array[O2] = new Array[O2](size)
-    copyToArray(array)
-    ArraySeq.unsafeWrapArray[O2](array)
-  }
-
-  def toArraySeqUntagged: ArraySeq[O] = {
-    val buf = ArraySeq.untagged.newBuilder[O]
-    buf.sizeHint(size)
-    var i = 0
-    while (i < size) {
-      buf += apply(i)
-      i += 1
-    }
-    buf.result()
-  }
-
+private[fs2] trait ChunkPlatform[+O] extends Chunk213And3Compat[O] { self: Chunk[O] =>
   def toIArray[O2 >: O: ClassTag]: IArray[O2] = IArray.unsafeFromArray(toArray)
 
   def toIArraySlice[O2 >: O](implicit ct: ClassTag[O2]): Chunk.IArraySlice[O2] =
@@ -56,27 +38,12 @@ private[fs2] trait ChunkPlatform[+O] { self: Chunk[O] =>
     }
 }
 
-private[fs2] trait ChunkCompanionPlatform { self: Chunk.type =>
+private[fs2] trait ChunkCompanionPlatform extends ChunkCompanion213And3Compat { self: Chunk.type =>
 
-  protected def platformIterable[O](i: Iterable[O]): Option[Chunk[O]] =
-    i match {
-      case a: immutable.ArraySeq[O] => Some(arraySeq(a))
-      case _                        => None
-    }
-
-  /** Creates a chunk backed by an immutable `ArraySeq`.
-    */
-  def arraySeq[O](arraySeq: immutable.ArraySeq[O]): Chunk[O] = {
-    val arr = arraySeq.unsafeArray.asInstanceOf[Array[O]]
-    array(arr)(ClassTag[O](arr.getClass.getComponentType))
-  }
-
-  /** Creates a chunk backed by an immutable array.
-    */
+  /** Creates a chunk backed by an immutable array. */
   def iarray[O: ClassTag](arr: IArray[O]): Chunk[O] = new IArraySlice(arr, 0, arr.length)
 
-  /** Creates a chunk backed by a slice of an immutable array.
-    */
+  /** Creates a chunk backed by a slice of an immutable array. */
   def iarray[O: ClassTag](arr: IArray[O], offset: Int, length: Int): Chunk[O] =
     new IArraySlice(arr, offset, length)
 
@@ -121,8 +88,4 @@ private[fs2] trait ChunkCompanionPlatform { self: Chunk.type =>
         ByteVector.view(values.asInstanceOf[Array[Byte]], offset, length)
       else ByteVector.viewAt(i => apply(i.toInt), size.toLong)
   }
-
-  /** Creates a chunk from a `scala.collection.IterableOnce`. */
-  def iterableOnce[O](i: collection.IterableOnce[O]): Chunk[O] =
-    iterator(i.iterator)
 }
